[refactor] Move pages into subfolders for cleaner organisation
